# docs/swagger.yaml
basePath: /
definitions:
  apistruct.ApiResponse:
    properties:
      data: {}
      errCode:
        example: 0
        type: integer
      errDlt:
        example: ""
        type: string
      errMsg:
        example: success
        type: string
    type: object
...
  apistruct.GetConfigListResp:
    properties:
      configNames:
        example:
        - log.yaml
        - kafka.yaml
        - redis.yaml
        items:
          type: string
        type: array
      environment:
        example: production
        type: string
      version:
        example: v3.8.3
        type: string
    type: object
  apistruct.GetConfigReq:
    properties:
      configName:
        example: log.yaml
        type: string
    required:
    - configName
    type: object
...
  apistruct.GetEnableConfigManagerResp:
    properties:
      enable:
        example: true
        type: boolean
    type: object
...
  apistruct.SetConfigReq:
    properties:
      configName:
        example: log.yaml
        type: string
      data:
        example: '{"level":"info","format":"json"}'
        type: string
    required:
    - configName
    - data
    type: object
  apistruct.SetEnableConfigManagerReq:
    properties:
      enable:
        example: true
        type: boolean
    type: object
...
host: localhost:10002
info:
  contact:
    email: contact@openim.io
    name: OpenIM Support
    url: https://github.com/openimsdk/open-im-server
  description: |-
    OpenIM Server 提供即时通讯服务的 REST API 接口文档

    ## 认证方式
    大部分接口需要在 Header 中携带 token 进行认证：
    - Header名称: token
    - 获取方式: 调用 /auth/get_user_token 接口

    ## 接口规范
    - 所有接口均使用 POST 方法
    - Content-Type: application/json
    - 响应格式: JSON
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenIM Server API
  version: "3.8"
paths:
...
  /config/get_config:
    post:
      consumes:
      - application/json
      description: 根据配置名称获取配置内容（需要管理员权限）
      parameters:
      - description: 配置名称
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apistruct.GetConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应，返回配置JSON字符串
          schema:
            allOf:
            - $ref: '#/definitions/apistruct.ApiResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      security:
      - BearerToken: []
      summary: 获取指定配置
      tags:
      - 配置管理
  /config/get_config_list:
    post:
      consumes:
      - application/json
      description: 获取所有可用的配置名称列表、环境信息和版本信息（需要管理员权限）
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/apistruct.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/apistruct.GetConfigListResp'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      security:
      - BearerToken: []
      summary: 获取配置列表
      tags:
      - 配置管理
  /config/get_enable_config_manager:
    post:
      consumes:
      - application/json
      description: 获取配置管理器的启用状态（需要管理员权限）
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/apistruct.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/apistruct.GetEnableConfigManagerResp'
              type: object
        "400":
          description: 获取状态失败
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      security:
      - BearerToken: []
      summary: 获取配置管理器状态
      tags:
      - 配置管理
  /config/reset_config:
    post:
      consumes:
      - application/json
      description: 重置所有配置为初始值（需要管理员权限，异步执行）
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应（异步执行）
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      security:
      - BearerToken: []
      summary: 重置配置
      tags:
      - 配置管理
  /config/set_config:
    post:
      consumes:
      - application/json
      description: 修改指定配置的内容（需要管理员权限，仅ETCD模式支持）
      parameters:
      - description: 配置信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apistruct.SetConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "400":
          description: 请求参数错误或非ETCD模式
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      security:
      - BearerToken: []
      summary: 设置单个配置
      tags:
      - 配置管理
  /config/set_enable_config_manager:
    post:
      consumes:
      - application/json
      description: 启用或禁用配置管理器（需要管理员权限，仅ETCD模式支持）
      parameters:
      - description: 启用状态
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apistruct.SetEnableConfigManagerReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "400":
          description: 请求参数错误或非ETCD模式
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      security:
      - BearerToken: []
      summary: 设置配置管理器状态
      tags:
      - 配置管理
...
  /restart:
    post:
      consumes:
      - application/json
      description: 通过ETCD通知所有服务实例重启（需要管理员权限，异步执行）
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应（异步执行）
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      security:
      - BearerToken: []
      summary: 重启服务
      tags:
      - 配置管理
...
schemes:
- http
- https
securityDefinitions:
  BearerToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"